requirements.txt
wolframalpha==5.0.0
python-telegram-bot==13.11
python-dotenv==0.19.2